+++ src/main/java/net/litetex/authback/mixin/server/DedicatedServerPropertiesMixin.java
@@ -4,7 +4,9 @@ import org.slf4j.Logger;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 
 import net.litetex.authback.server.AuthBackServer;
 import net.litetex.authback.shared.mixin.log.MixinLogger;
@@ -17,18 +19,22 @@ public abstract class DedicatedServerPropertiesMixin
 	@Unique
 	private static final Logger LOG = MixinLogger.server("DedicatedServerPropertiesMixin");
 	
-	@Redirect(
+	@WrapOperation(
 		method = "<init>",
 		at = @At(
 			value = "INVOKE",
 			target = "Lnet/minecraft/server/dedicated/DedicatedServerProperties;get(Ljava/lang/String;Z)Z")
 	)
-	boolean get(final DedicatedServerProperties instance, final String key, final boolean defaultValue)
+	boolean get(
+		final DedicatedServerProperties instance,
+		final String key,
+		final boolean defaultValue,
+		final Operation<Boolean> original)
 	{
-		final boolean resolvedValue = instance.get(key, defaultValue);
+		final boolean resolvedValue = original.call(instance, key, defaultValue);
 		if("enforce-secure-profile".equals(key)
 			&& resolvedValue
-			&& AuthBackServer.instance().isForceDisableEnforceSecureProfile())
+			&& AuthBackServer.instance().config().forceDisableEnforceSecureProfile())
 		{
 			LOG.info("enforce-secure-profile was enabled - force disabling it. "
 				+ "You should disable it in the server.properties!");